+++ pkg/agent/collectors/events/service.go
@@ -10,11 +10,11 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/log"
 
 	"github.com/giantswarm/observability-operator/pkg/agent"
-	"github.com/giantswarm/observability-operator/pkg/auth"
 	"github.com/giantswarm/observability-operator/pkg/common/labels"
 	"github.com/giantswarm/observability-operator/pkg/common/organization"
 	"github.com/giantswarm/observability-operator/pkg/common/tenancy"
 	"github.com/giantswarm/observability-operator/pkg/config"
+	"github.com/giantswarm/observability-operator/pkg/credential"
 )
 
 const (
@@ -27,12 +27,9 @@ type Service struct {
 	ConfigurationRepository agent.ConfigurationRepository
 	OrganizationRepository  organization.OrganizationRepository
 	TenantRepository        tenancy.TenantRepository
-	LogsAuthManager         auth.AuthManager
-	TracesAuthManager       auth.AuthManager
-	MetricsAuthManager      auth.AuthManager
 }
 
-func (s *Service) ReconcileCreate(ctx context.Context, cluster *clusterv1.Cluster, observabilityBundleVersion semver.Version, caBundle string) error {
+func (s *Service) ReconcileCreate(ctx context.Context, cluster *clusterv1.Cluster, observabilityBundleVersion semver.Version, caBundle string, creds credential.BackendCredentials) error {
 	logger := log.FromContext(ctx)
 	logger.Info("alloy-events-service - ensuring alloy events is configured")
 
@@ -45,7 +42,7 @@ func (s *Service) ReconcileCreate(ctx context.Context, cluster *clusterv1.Cluste
 		return fmt.Errorf("failed to generate alloy events configmap: %w", err)
 	}
 
-	secretData, err := s.GenerateAlloyEventsSecretData(ctx, cluster, loggingEnabled, tracingEnabled, monitoringEnabled, caBundle)
+	secretData, err := s.GenerateAlloyEventsSecretData(cluster, loggingEnabled, tracingEnabled, monitoringEnabled, caBundle, creds)
 	if err != nil {
 		return fmt.Errorf("failed to generate alloy events secret: %w", err)
 	}