+++ pkg/agent/collectors/logs/service.go
@@ -9,11 +9,11 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/log"
 
 	"github.com/giantswarm/observability-operator/pkg/agent"
-	"github.com/giantswarm/observability-operator/pkg/auth"
 	"github.com/giantswarm/observability-operator/pkg/common/labels"
 	"github.com/giantswarm/observability-operator/pkg/common/organization"
 	"github.com/giantswarm/observability-operator/pkg/common/tenancy"
 	"github.com/giantswarm/observability-operator/pkg/config"
+	"github.com/giantswarm/observability-operator/pkg/credential"
 )
 
 const (
@@ -41,10 +41,9 @@ type Service struct {
 	ConfigurationRepository agent.ConfigurationRepository
 	OrganizationRepository  organization.OrganizationRepository
 	TenantRepository        tenancy.TenantRepository
-	LogsAuthManager         auth.AuthManager
 }
 
-func (s *Service) ReconcileCreate(ctx context.Context, cluster *clusterv1.Cluster, observabilityBundleVersion semver.Version, caBundle string) error {
+func (s *Service) ReconcileCreate(ctx context.Context, cluster *clusterv1.Cluster, observabilityBundleVersion semver.Version, caBundle string, creds credential.BackendCredentials) error {
 	logger := log.FromContext(ctx)
 	logger.Info("alloy-logs-service - ensuring alloy logs is configured")
 
@@ -56,7 +55,7 @@ func (s *Service) ReconcileCreate(ctx context.Context, cluster *clusterv1.Cluste
 		return fmt.Errorf("failed to generate alloy logs configmap: %w", err)
 	}
 
-	secretData, err := s.GenerateAlloyLogsSecretData(ctx, cluster, loggingEnabled, caBundle)
+	secretData, err := s.GenerateAlloyLogsSecretData(cluster, loggingEnabled, caBundle, creds)
 	if err != nil {
 		return fmt.Errorf("failed to generate alloy logs secret: %w", err)
 	}